Introduced MODE_SPI_INC_8A
This changeset introduces support for high resolution, incremental SPI
encoders.

The actual encoder that we are using is a Renesas IPS2200, but the
analog signal is being processed by a 24-bit external ADC
(MCP3562RT-E/NC) that has an SPI output.

The existing SPI workflow is being used to pull the values into DMA, but
we had to update a variety of variable names to indicate support for
incremental and not just absolute values.

We also needed to update the buffers to support 32 bit values, rather
than just 16 bits.